refactor: move method to test_utils
