+++ crates/web-client/src/utils/test_utils.rs
@@ -1,8 +1,14 @@
-use miden_client::testing::account_id::ACCOUNT_ID_REGULAR_PRIVATE_ACCOUNT_UPDATABLE_CODE;
+use miden_client::{
+    testing::account_id::ACCOUNT_ID_REGULAR_PRIVATE_ACCOUNT_UPDATABLE_CODE,
+    transaction::TransactionResult as NativeTransactionResult,
+};
 use miden_objects::account::AccountId as NativeAccountId;
 use wasm_bindgen::prelude::*;
 
-use crate::models::account_id::AccountId;
+use crate::{
+    WebClient,
+    models::{account_id::AccountId, transaction_result::TransactionResult},
+};
 
 #[wasm_bindgen]
 pub struct TestUtils;
@@ -14,5 +20,29 @@ impl TestUtils {
         let native_account_id: NativeAccountId =
             ACCOUNT_ID_REGULAR_PRIVATE_ACCOUNT_UPDATABLE_CODE.try_into().unwrap();
         native_account_id.into()
+    }
+}
+// WEB CLIENT TESTING HELPERS
+// ================================================================================================
+
+#[cfg(feature = "testing")]
+#[wasm_bindgen]
+impl WebClient {
+    #[wasm_bindgen(js_name = "testingApplyTransaction")]
+    pub async fn testing_apply_transaction(
+        &mut self,
+        tx_result: TransactionResult,
+    ) -> Result<(), JsValue> {
+        let native_transaction_result: NativeTransactionResult = tx_result.into();
+
+        if let Some(client) = self.get_mut_inner() {
+            client
+                .testing_apply_transaction(native_transaction_result)
+                .await
+                .map_err(|err| JsValue::from_str(&format!("Failed to apply transaction: {err}")))?;
+            Ok(())
+        } else {
+            Err(JsValue::from_str("Client not initialized"))
+        }
     }
 }
